requirements.txt
# All packages below are already installed in the env_vllm conda env.
# This file is a reference manifest only; no pip install is required.
vllm>=0.19.0
openai>=2.0
transformers>=4.57
torch>=2.10
opencv-python-headless>=4.13
pillow>=10.0
numpy>=1.26
